+++ src/types.h
@@ -10,8 +10,8 @@
 /* Some generic, cross-platform types. */
 
 struct _MMPoint {
-	size_t x;
-	size_t y;
+	int32_t x;
+	int32_t y;
 };
 
 typedef struct _MMPoint MMPoint;
@@ -25,8 +25,8 @@ struct _MMSignedPoint {
 typedef struct _MMSignedPoint MMSignedPoint;
 
 struct _MMSize {
-	size_t width;
-	size_t height;
+	int32_t width;
+	int32_t height;
 };
 
 typedef struct _MMSize MMSize;
@@ -38,7 +38,7 @@ struct _MMRect {
 
 typedef struct _MMRect MMRect;
 
-H_INLINE MMPoint MMPointMake(size_t x, size_t y)
+H_INLINE MMPoint MMPointMake(int32_t x, int32_t y)
 {
 	MMPoint point;
 	point.x = x;
@@ -54,7 +54,7 @@ H_INLINE MMSignedPoint MMSignedPointMake(int32_t x, int32_t y)
 	return point;
 }
 
-H_INLINE MMSize MMSizeMake(size_t width, size_t height)
+H_INLINE MMSize MMSizeMake(int32_t width, int32_t height)
 {
 	MMSize size;
 	size.width = width;
@@ -62,7 +62,7 @@ H_INLINE MMSize MMSizeMake(size_t width, size_t height)
 	return size;
 }
 
-H_INLINE MMRect MMRectMake(size_t x, size_t y, size_t width, size_t height)
+H_INLINE MMRect MMRectMake(int32_t x, int32_t y, int32_t width, int32_t height)
 {
 	MMRect rect;
 	rect.origin = MMPointMake(x, y);
@@ -75,14 +75,14 @@ H_INLINE MMRect MMRectMake(size_t x, size_t y, size_t width, size_t height)
 #if defined(IS_MACOSX)
 
 #define CGPointFromMMPoint(p) CGPointMake((CGFloat)(p).x, (CGFloat)(p).y)
-#define MMPointFromCGPoint(p) MMPointMake((size_t)(p).x, (size_t)(p).y)
+#define MMPointFromCGPoint(p) MMPointMake((int32_t)(p).x, (int32_t)(p).y)
 
 #define CGPointFromMMSignedPoint(p) CGPointMake((CGFloat)(p).x, (CGFloat)(p).y)
 #define MMSignedPointFromCGPoint(p) MMPointMake((int32_t)(p).x, (int32_t)(p).y)
 
 #elif defined(IS_WINDOWS)
 
-#define MMPointFromPOINT(p) MMPointMake((size_t)p.x, (size_t)p.y)
+#define MMPointFromPOINT(p) MMPointMake((int32_t)p.x, (int32_t)p.y)
 
 #endif
